third_party/blink/web_tests/fast/dom/HTMLProgressElement/progress-element-with-child-crash-expected.txt

CONSOLE ERROR: A descendant of a <select> does not follow the content model.
Test for Bug 48019. It is OK not to crash.
